requirements.txt
Flask==3.1.2
Pillow
exifread
PyPDF2
python-docx
mutagen
filetype
